Revert Coverity Scan to build within the source
This is not pretty, but so far as I can tell, Coverity Scan must be run
within the git clone (or it reports "fatal: Not a git repository", and
fails to gather SCM data).

Also moved the cppcheck run to the before_script section so it runs on
the git clone code prior to Coverity Scan polluting it.

If we can find a (supported) way to run Coverity Scan out of the git
clone, then we can undo this change.