Use explicit types for pcp::units function arguments
Templating the type was overkill, given that the `PMDA_PMUNITS` macro
only exists to automatically order the arguments correctly for the
`pmUnits` member of the `pmDesc` structure (the order reverses on some
platforms). But also, the templating was forcing all parameters to be
the same type, whereas the `pmUnits` structure members are not all the
same type.

Given that `pmUnits` has not changed in 8 years, it makes most sense to
replace the templating with the explicit `pmUnits` types directly.